+++ src/Client/ClientCore/Systems/Input/InGameKeyboardShortcuts.cs
@@ -48,6 +48,7 @@ public class InGameKeyboardShortcuts
         shortcutTable[SDL.SDL_Keycode.SDLK_ESCAPE] = () => Toggle(ClientStateFlag.ShowInGameMenu);
         shortcutTable[SDL.SDL_Keycode.SDLK_F2] = () => Toggle(ClientStateFlag.ShowPlayerDebug);
         shortcutTable[SDL.SDL_Keycode.SDLK_F3] = () => Toggle(ClientStateFlag.ShowEntityDebug);
+        shortcutTable[SDL.SDL_Keycode.SDLK_F4] = () => Toggle(ClientStateFlag.ShowRendererDebug);
         shortcutTable[SDL.SDL_Keycode.SDLK_INSERT] = () => Toggle(ClientStateFlag.ShowTemplateEntityEditor);
         shortcutTable[SDL.SDL_Keycode.SDLK_DELETE] = () =>
         {