+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### 23 March 2026
 
 * Fix bug where block entities were directly persisted to the database. Blocks are only persisted indirectly through the world segment block data table. Since block IDs are treated as volatile and are reused between server runs, this leads to crashes caused by duplicate entity IDs.
+* Add renderer debug GUI (press F4 to open).
 
 #### 22 March 2026